Added a lot of changes. Read the description.
The DismembermentMissionLogic and Dismemberment classes have been united;
The static modifiers have been removed from the all methods;
The SpawnAgent method has been moved to the TOWDebug class;
The KillAgent method has been removed;
The calls of the DisplayDebugMessage method have been replaced by TOWCommon.Say();
The DisplayDebugMessage has been removed;
The isCompliant variable has been renamed to canBeDismembered;
The Settings.cfg file has been removed.
