--- Settings.cfg
@@ -1,5 +1,0 @@
-// Mount & Blade II Bannerlord - Dismemberment Mod by jedijosh920
-// Settings configuration file
-Chance=1
-SlowMotion=false
-Debug=true
